+++ EnvironmentSimulator/Modules/RoadManager/inc/Line.hpp
@@ -1,5 +1,6 @@
 #ifndef LINE_HPP
 #define LINE_HPP
+#include "CommonMini.hpp"
 #include "Geometry.hpp"
 #include "pugixml.hpp"
