+++ EnvironmentSimulator/Modules/RoadManager/inc/ParamPoly3.hpp
@@ -1,6 +1,7 @@
 #ifndef PARAMPOLY3_HPP
 #define PARAMPOLY3_HPP
 
+#include "CommonMini.hpp"
 #include "Geometry.hpp"
 #include "pugixml.hpp"
