+++ EnvironmentSimulator/Modules/RoadManager/inc/RoadObject.hpp
@@ -16,7 +16,7 @@ class RoadObject {
 	void AddUserData(std::shared_ptr<UserData> userData) { user_data_.push_back(userData); }
 	std::vector<ValidityRecord> validity_;
 
-   private:
+   protected:
 	std::vector<std::shared_ptr<UserData>> user_data_;
 };
